unet-segmentation/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.1
opencv-python>=4.9.0
numpy>=1.26.0,<2.0
pandas>=2.2.0
matplotlib>=3.8.0
pillow>=10.2.0
pyyaml>=6.0.1
pytest>=8.0.0
pytest-cov>=4.1.0
tensorboard>=2.16.0
